+++ config/config.exs
@@ -25,6 +25,8 @@ config :logger, :console,
 # Use Jason for JSON parsing in Phoenix
 config :phoenix, :json_library, Jason
 
+config :phoenix, :template_engines, md: PhoenixMarkdown.Engine
+
 # Admin
 
 config :ex_admin,