+++ lib/birdiy_web/views/privacy_view.ex
@@ -1,0 +1,3 @@
+defmodule BirdiyWeb.PrivacyView do
+  use BirdiyWeb, :view
+end