+++ lib/birdiy_web/views/terms_view.ex
@@ -1,0 +1,3 @@
+defmodule BirdiyWeb.TermsView do
+  use BirdiyWeb, :view
+end